Move numpy-related functions into test_config.py. Replace has_numpy flag with pytest.importorskip(). Add numpy to requirements.
