+++ requirements.txt
@@ -5,3 +5,4 @@ srsly>=2.4.0,<3.0.0
 pytest>=5.2.0,!=7.1.0
 mypy>=0.901,<0.970; platform_machine!='aarch64'
 types-dataclasses>=0.1.3; python_version < "3.7"
+numpy>=1.15.0
